requirements.txt
aiohttp==3.5.4
ares==0.5
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Login==0.2.11
Flask-PyMongo==2.3.0
idna==2.8
idna-ssl==1.1.0
importlib-metadata==0.18
inflect==2.1.0
irc==17.1
itsdangerous==1.1.0
jaraco.classes==2.0
jaraco.collections==2.0
jaraco.functools==2.0
jaraco.itertools==4.4.2
jaraco.logging==2.0
jaraco.stream==2.0
jaraco.text==3.0
Jinja2==2.10.1
lxml==4.3.4
MarkupSafe==1.1.1
more-itertools==7.1.0
multidict==4.5.2
passlib==1.7.1
pycvesearch==0.5
pymongo==3.8.0
python-dateutil==2.8.0
pytz==2019.1
redis==3.2.1
requests==2.22.0
six==1.12.0
slackclient==2.1.0
sleekxmpp==1.3.3
tempora==1.14.1
termcolor==1.1.0
tornado==6.0.3
typing-extensions==3.7.4
ujson==1.35
urllib3==1.25.3
Werkzeug==0.15.4
Whoosh==2.7.4
xlrd==1.2.0
yarl==1.3.0
zipp==0.5.2
